LayoutTests/imported/w3c/web-platform-tests/svg/animations/scripted/animatetransform-type-missing-value-default-expected.txt


PASS <animateTransform> 'type' attribute missing/invalid value default, "type" attribute is "translate"
FAIL <animateTransform> 'type' attribute missing/invalid value default, missing "type" attribute assert_equals: expected 30 but got 0
PASS <animateTransform> 'type' attribute missing/invalid value default, invalid "type" attribute
FAIL <animateTransform> 'type' attribute missing/invalid value default, removed "type" attribute assert_equals: expected 70 but got 0
